refactor settings page using react context
