+++ client/src/pages/Settings.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useReducer } from "react";
+import React, { useState, useEffect, useReducer, useMemo } from "react";
 import { makeStyles, Grid } from "@material-ui/core";
 
 import axios from "axios";
@@ -11,7 +11,6 @@ import {
   settingsReducer,
   settingsInitialState,
 } from "../utils/settings-context";
-import { getUser, storeUser } from "../utils/localStorage";
 
 const useStyles = makeStyles((theme) => ({
   container: {
@@ -25,16 +24,16 @@ const useStyles = makeStyles((theme) => ({
 const Settings = (props) => {
   const classes = useStyles();
   const [currentTabIndex, setTabIndex] = useState(0);
-  const [userCompanies, setUserCompanies] = useState([]);
-  const [user, setUser] = useState(getUser());
 
   const [state, dispatch] = useReducer(settingsReducer, settingsInitialState);
+  const context = useMemo(() => {
+    return { state, dispatch };
+  }, [state, dispatch]);
 
   // api fetch
   useEffect(() => {
     async function getCompanies() {
       const res = await axios.get("/api/company");
-      // setUserCompanies(res.data.companies);
       dispatch({ type: "get_companies", payload: res.data.companies });
     }
     getCompanies();
@@ -44,37 +43,14 @@ const Settings = (props) => {
     setTabIndex(newValue);
   };
 
-  const removeCompanyFromUser = async (companyName) => {
-    // note the delete method is the only method where if you require request body data
-    // you have to do it in this way, other methods let you pass in a javascript object directly
-    // without having to preface with data
-    await axios.delete("/api/company", { data: { name: companyName } });
-    setUserCompanies(
-      userCompanies.filter((company) => company.name !== companyName)
-    );
-  };
-
-  const updateSubscriberEmail = async (email) => {
-    await axios.put("/api/users/subscribe-mail/update", {
-      subscriberEmail: email,
-    });
-    // update user object on client
-    storeUser({ ...user, subscriberEmail: email });
-  };
-
   return (
-    <SettingsContext.Provider value={{ state, dispatch }}>
+    <SettingsContext.Provider value={context}>
       <Grid container className={classes.container}>
         <SidePanel
           setTabIndex={onChangeTabs}
           currentTabIndex={currentTabIndex}
         />
-        <SettingsContent
-          currentTabIndex={currentTabIndex}
-          user={user}
-          removeCompany={removeCompanyFromUser}
-          updateSubEmail={updateSubscriberEmail}
-        />
+        <SettingsContent currentTabIndex={currentTabIndex} />
       </Grid>
     </SettingsContext.Provider>
   );
